pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["None"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.0"
python-jose = "^3.3.0"
SQLAlchemy = "1.4.26"
uvicorn = "^0.14.0"
asyncpg = "^0.25.0"
alembic = "^1.7.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
